Fixed ignore rule
The intention was to ignore every module except those prefixed with "nl_".
However, the modules/ rule made it so that the exception (!modules/nl_*) did not
work. This is because the modules directory was being ignored at the top level.
By bringing the ignore rule down a level (to ignoring everything INSIDE the
modules directory as opposed to the modules directory itself), the exception
rule works. Now nl_ modules are tracked.

The short of it is, you don't want to ignore the directory because you want to
recurse INTO them. That way, the exception will work.

Signed-off-by: Byron Sanchez <byron@hackbytes.com>