Added an nginx types configuration file
The types configuration files is used to define how certain file types are 
served. In this case, asc files are going to be treated as plain-text public key 
files. Therefore, they must be served as plain-text files. Without this 
particular configuration, the asc files will not be served as plain text (and 
viewable in the browser) and instead, will be offered as a file download.