Update deployment strategy structure
Deployment code will be in it's own directory.

Repository mirroring will leverage this deployment strategy to automate
mirroring by using path/naming conventions. The mirror scripts will
expect a certain path containing the list of git repositories to mirror
to, and if it finds it, it'll push to those repositories.