Add git-ping script
This script will regularly check the RSS feeds for all git repos. When
changes are detected, all external repos will be updated (github,
bitbucket, etc.).

This is great because a cron is set to check each minute, without
actually pushing every minute. It'll only push when actual changes are
detected.